pyproject.toml
[project]
name = "enterprise-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "jieba>=0.42.1",
    "openai>=2.31.0",
    "pdfplumber>=0.11.9",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "rich>=15.0.0",
    "sentence-transformers>=5.4.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
